Major Changes:
The consumeArea algorithm now restarts every time it finds a point it can consume. This is relatively very time consuming, but it ensures that all 4 consumption directions (e.g. N/E/S/W) have been consumed before moving on to the next consumable pixel. This makes it significantly less likely that large 'tendrils' will form where a pixel is added, the north is added, the north of that is added, and so on, until we have a chain of 10 pixels north and no other expansion.To restate, this change makes the algorithm prefer to consume points that are as close as possible to points that have already been consumed. This also makes concave polygons much less probable, and convex polygons seem to work better with the other algorithms (shrink, marge, and even consumeArea with the same poly).

The isConsumable algorithm is more forgiving if there are 2 neighboring pixels. If there were 2 previously, then it would report that pixel as not consumable. However, now it checks for 'corners', where the sampled pixel has a N/E, a N/W, a S/E, or a S/W combination available. In these cases, it checks if the other 'corner' of the 2x2 square exists. If it does, then consuming the sampled pixel will not make the polygon non-contigious, and therefore isConsumable reports true. This makes the polygons consume more naturally, whereas before this change they would occasionally find a corner that could not be consumed, so they would start consuming at the closest available location, which was typically a distance of one pixel from the not-consumable 'corner' point. This error would propagate as the consumption continued: the corner would never be consumed because isConsumable continually reported it as not consumable, and a one pixel tendril would emerge from the body of the polygon being consumed to the corner that was not consumable, due to consumeArea's desire to not create non-contigious polygons.

The MergeRanking was introduced. This orders a list, placing the points that are most optimal 'starting locations' for merges at the beginning of the list. Naturally this is dependent upon the internal implementation of the merge algorithm (e.g. consumeArea), but for now this operates by finding all 'leaf' points, e.g. points that only have one neighbor, and placing them at the front of the list. This attempts to ensure that the merge starts on the interior of any concave regions within this polygon. It's not 100% successful at this, but it does make a good attempt. I need to make getStartingPoint use this ranking method, and then I need to remove the need for consumeArea to be passed a starting point - we can just call getStartingPoint internally.