Removed the 'isContigious' sanity check
Turns out that it only checks if there is a single pixel that is
out in left field. If the poly is not conigious, but each separate area
has at least two pixels, then this method will return that the overall
poly is in fact contigious (an error). This largely makes this useless,
and we need to replace it with a method that ensures that every pixel in
the poly can be reached from some arbitrary starting pixel.