Added day timeslice, added 'jump ahead' in data file, better docs
The jump ahead essentially seeks the input data file by large amounts until
it finds a range that falls within the desired filters, and then it resets
itself to the last marked point before it found this range.

The day timeslice was added, as it's useful.

The documentation was updated to better reflect what the filters do (as
the year/month/day filters do something different than the time filter).
This filter impl was chosen mainly so that I could compare against other
datasets.